Handle joining to an existing cluster
Only add the local instance URL when reconciling.

It's not possible to add more than one etcd node to the member list at a
time with small clusters. During the time between adding the member and
the member starting up, no changes can be made in etcd due to lack of
quorum. This causes etcd-bootstrap to fail when adding more than 1 node
to an existing 1 node cluster.

By only adding the local instance we can bypass the problem - the local
node if it gets there first will be able to make progress and start
itself. Other nodes will retry until they can join the cluster.

Only include existing nodes plus new node.

If we include all nodes when joining, etcd will complain about a member
mismatch. We have to include only the joining node along with existing
live cluster members.

Ignore if removal fails.

This can happen when there is not enough quorum to remove the invalid
node - which leads to a broken state where etcd-bootstrap can't generate
the configuration. Instead ignore it and allow etcd to continue to start
up.