Allow bootstrapping functionality to be invoked without starting etcd
This change allows etcd-bootstrap to be used with coreos' etcd images
from v3 onwards. This is achieved by having etcd-bootstrap output a
startup script into a volume which can then be mounted into the coreos
image and the entrypoint changed to the generated startup script.

This may be utilised in Kubernetes by running etcd-bootstrap as an
init-container, and then sharing the startup script in an emptyDir
volume.

- Splits `start-etcd.sh` into `bootstrap.sh` and `etcd.sh` to allow the
two changes to allow bootstrapping without starting etcd and starting
etcd without bootstrapping.
- Changed location of intermediary config file to /bootstrap to make
volume mounting easier.
- Move etcd binary location to ``/usr/local/bin` from `/` to match the
coreos image