Initial commit for refactor
Add registration types and tests

We will now support the concept of registration types which will, at the
start be either dns or lb (none to disable). This is to refactor to
support this and also to add support for lb registration for aws.

Add lb registration type and test.

Update README and CHANGELOG to release v1.3.0.

Add elbv2 vendor package