Refactor etcd-bootstrap
This is to make code design changes to etcd-boostrap such that the
project can be better maintained and tests. Changes include:
* Using cobra for command line arguments
* Split out provider from bootstrap logic to allow for better code
segregation
* Support mocking for the etcd member client in better test code
coverage
* More extensive testing for both the bootstrap and AWS packages
* Replace vendor packages with go modules
* Versioned binaries (can be printed using `./etcd-boostrap --version`)

It also includes adding support for multiple registration types
(currently on AWS provider supports this) where you can choose to
register the cluster information with either a route53 DNS record or an
AWS target group.