Refactor etcd-bootstrap and add support for AWS target groups (#27)
* Remove vendor packages

* Refactor etcd-bootstrap

This is to make code design changes to etcd-boostrap such that the
project can be better maintained and tests. Changes include:
* Using cobra for command line arguments
* Split out provider from bootstrap logic to allow for better code
segregation
* Support mocking for the etcd member client in better test code
coverage
* More extensive testing for both the bootstrap and AWS packages
* Replace vendor packages with go modules
* Versioned binaries (can be printed using `./etcd-boostrap --version`)

It also includes adding support for multiple registration types
(currently on AWS provider supports this) where you can choose to
register the cluster information with either a route53 DNS record or an
AWS target group.

* Add tests for refactored code

This is to move over some of the tests from the code that was refactored
and also to add more tests for better code test coverage and also better
readibiltiy.

* Update README and CHANGELOG

This is to release v2.0.0

Also includes some changes to the README to reflect the new
etcd-bootstrap flags and user interaction.

* Upgrade go travis version to 1.12

* Adjust Makefile to support a single container

We have decided to remove the etcd wrapper container and have
etcd-bootstrap run as an init-container. The build pipeline will
now create a container than has a built version of etcd-boostrap
and will not contain a downloaded version of etcd too.

* Add wrapper scripts

When running etcd-bootstrap as a init container we need a way of
generating the ENV vars and then allowing the vanilla etcd container to
source them. The provided scripts allow you to do this.