Add SRV support for looking up cluster instances
This adds new flags to the AWS bootstrapper:
- `clusterLookupMethod=srv` enables SRV lookup
- `srvDomainName=my-cluster.example.com` specifies the domain to find
the SRV record on
- `srvService=etcd-cluster-ssl` specifies the SRV service to use

This follows RFC4122, so will look up records in
`_etcd-cluster-ssl._tcp.my-cluster.example.com`. Each target in the SRV
record is added as an instance for cluster bootstrapping and discovery.
Currently all fields other than `target` are ignored in the SRV record.