SRV support for etcd-bootstrap (#33)
This adds new flags to the AWS bootstrapper:
- `instance-lookup-method=srv` enables SRV lookup
- `srv-domain-name=my-cluster.example.com` specifies the domain to find
the SRV record on
- `srv-service=etcd-cluster-ssl` specifies the SRV service to use

This follows RFC4122, so will look up records in
`_etcd-cluster-ssl._tcp.my-cluster.example.com`. Each target in the SRV
record is added as an instance for cluster bootstrapping and discovery.
Currently all fields other than `target` are ignored in the SRV record.

I had to change the structure of the `GetInstances()` and `GetLocalInstance()` methods to return an error, so I could move the AWS cloud logic into these methods rather than calculating them in `New`. This made it possible to reuse the AWS logic for getting the local instance information still.