Fix advertise URLs being different to the local IP (#38)
This is a fairly significant change as a lot of the internal
reconciliation logic of etcd-bootstrap was assuming local IP to be
equivalent to the advertise URL.

There were also some subtle edge cases that were not very obvious in the
code - I have tried to add additional comments, and improve the tests to
make these more obvious.

The most complex case is that where a peerURL has been added, but the
etcd node has not registered itself yet. We have to be careful to set
the initial cluster correctly - it should not include the to-be-added
node. See https://etcd.io/docs/v3.4.0/op-guide/runtime-configuration/
for details.